+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: oozsertt <oozsertt@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/14 17:42:58 by user42            #+#    #+#             */
-/*   Updated: 2020/09/17 15:19:57 by oozsertt         ###   ########.fr       */
+/*   Updated: 2020/09/18 01:02:54 by oozsertt         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
-	int i;
-	int j;
+	int		i;
+	size_t	j;
 
 	i = 0;
 	while (dst[i] != '\0')
